+++ supvisors/test/scripts/check_start_sequence.py
@@ -206,9 +206,9 @@ class CheckStartSequenceTest(CheckSequenceTest):
         """ Check the starting of the hmi program. """
         # define the expected events for the hmi program
         program = self.context.get_program('my_movies:hmi')
-        address = self.HOST_02 if self.HOST_02 in self.nodes else self.HOST_01
-        program.add_event(ProcessStateEvent(ProcessStates.STARTING, address))
-        program.add_event(ProcessStateEvent(ProcessStates.RUNNING, address))
+        node_name = self.HOST_02 if self.HOST_02 in self.nodes else self.HOST_01
+        program.add_event(ProcessStateEvent(ProcessStates.STARTING, node_name))
+        program.add_event(ProcessStateEvent(ProcessStates.RUNNING, node_name))
         # check that the events received are compliant
         self.check_events('my_movies')
         self.assertFalse(self.context.has_events('my_movies'))
@@ -230,9 +230,10 @@ class CheckStartSequenceTest(CheckSequenceTest):
         """ Check the starting of the test_reader program. """
         # define the expected events for the test_reader program
         program = self.context.get_program('player:test_reader')
-        program.add_event(ProcessStateEvent(ProcessStates.STARTING, self.HOST_01))
-        program.add_event(ProcessStateEvent(ProcessStates.RUNNING, self.HOST_01))
-        program.add_event(ProcessStateEvent(ProcessStates.EXITED, self.HOST_01))
+        node_name = self.HOST_03 if self.HOST_03 in self.nodes else self.HOST_01
+        program.add_event(ProcessStateEvent(ProcessStates.STARTING, node_name))
+        program.add_event(ProcessStateEvent(ProcessStates.RUNNING, node_name))
+        program.add_event(ProcessStateEvent(ProcessStates.EXITED, node_name))
         # check that the events received are compliant
         self.check_events('player')
         self.assertFalse(self.context.has_events('player'))